ADD: deserializer functionnal

+++ include/ECS/Components/TransformComponent.hpp
@@ -24,16 +24,16 @@ class TransformComponent : public IComponent
     TransformComponent& operator=(const TransformComponent& rhs) noexcept = delete;
     TransformComponent& operator=(TransformComponent&& rhs) noexcept      = delete;
 
-    int  getX() const noexcept;
-    int  getY() const noexcept;
-    void setX(int x) noexcept;
-    void setY(int y) noexcept;
-    void setPos(int x, int y) noexcept;
-    int  getScaleX() const noexcept;
-    int  getScaleY() const noexcept;
-    void setScaleX(float x) noexcept;
-    void setScaleY(float y) noexcept;
-    void setScale(float x, float y) noexcept;
+    int   getX() const noexcept;
+    int   getY() const noexcept;
+    void  setX(int x) noexcept;
+    void  setY(int y) noexcept;
+    void  setPos(int x, int y) noexcept;
+    float getScaleX() const noexcept;
+    float getScaleY() const noexcept;
+    void  setScaleX(float x) noexcept;
+    void  setScaleY(float y) noexcept;
+    void  setScale(float x, float y) noexcept;
 
   private:
     int   _x;
